pyproject.toml
[tool.poetry]
name = "fastapi-was"
version = "0.1.0"
description = ""
authors = ["miintto <minjp93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
uvicorn = "^0.15.0"
pydantic = "^1.8.2"
fastapi = "^0.70.0"
SQLAlchemy = "^1.4.27"
psycopg2 = "^2.9.2"
PyJWT = "^2.3.0"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
